vp: core: syscall: rework syscall/brk (fixed RV32 brk bug)
The brk syscall is used to get or modify the current heap pointer.
In any case, it returns the heap pointer, which is an unsigned
XLEN wide address (RV32 -> 32bit / RV64 -> 64 bit).

On RV32, the brk implementation contained a lexical cast to the (signed)
host "int" type, which fails if the heap pointer grows beyond 0x7ffff_ffff
(would be a negative values).
-> This is a bug!

Fix: Rework sys_brk, execut_syscall, ...
 * Remove the lexcial casts to signed host integer (address is unsigned)
 * Use uint64_t instead of host "int" (explicitly use/specify unsigned
   integers with width >= XLEN)

Additional changes:
 * Cleanup sys_brk implementation.
 * Added TODOs (future work)

+++ vp/src/core/rv64/syscall.h
@@ -146,7 +146,7 @@ struct SyscallHandler : public sc_core::sc_module, syscall_emulator_if {
 	 * host as byte array). Note: the data structures on the host system might
 	 * not be binary compatible with those on the guest system.
 	 */
-	int execute_syscall(uint64_t n, uint64_t _a0, uint64_t _a1, uint64_t _a2, uint64_t _a3);
+	uint64_t execute_syscall(uint64_t n, uint64_t _a0, uint64_t _a1, uint64_t _a2, uint64_t _a3);
 };
 
 }  // namespace rv64
